Re-factored helper functions outside of Stars Component
